DDF: all externals are now in ./base
Pushed new branch [experimental]; DDF folders are now in a root folder called /base, with the games renamed from [gamename]_ddf -> [gamename]. I did not copy this to the main branch because I have not finished testing yet, but so far it seems to work fine, and definitely cleans up the root EDGE directory.

+++ epi/utility.cc
@@ -89,7 +89,7 @@ namespace epi
     	
     	// Reset the pointers in the table to look at new data
     	int i;
-    	int offset = data - src.data;
+    	size_t offset = data - src.data;
     	for (i=0; i<numstrs; i++)
     	{
     		if (strs[i])
